+++ src/TestableHttpClient/Responses.cs
@@ -55,6 +55,7 @@ public static class Responses
     /// Create a response with the NoContent status code.
     /// </summary>
     /// <returns>An HttpResponse with the configured StatusCode.</returns>
+    [Obsolete("Please use StatusCode(HttpStatusCode.NoContent) instead.")]
     public static IResponse NoContent() => StatusCode(HttpStatusCode.NoContent);
     /// <summary>
     /// Create a response with some text content.
@@ -79,6 +80,23 @@ public static class Responses
     /// <param name="contentType">The content type of the response, defaults to 'application/json'.</param>
     /// <returns>A response with specific content.</returns>
     public static IResponse Json(object? content, HttpStatusCode statusCode, string? contentType = null, JsonSerializerOptions? jsonSerializerOptions = null) => new JsonResponse(content, contentType) { StatusCode = statusCode, JsonSerializerOptions = jsonSerializerOptions };
+    /// <summary>
+    /// Create a response for several routes.
+    /// </summary>
+    /// <param name="builder">The route builder that can be used to configure multiple routes.</param>
+    /// <returns>A response with routing capabilities.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when the builder paramater is null.</exception>
+    public static IResponse Route(Action<IRoutingResponseBuilder> builder)
+    {
+        if (builder is null)
+        {
+            throw new ArgumentNullException(nameof(builder));
+        }
+
+        RoutingResponseBuilder routingResponseBuilder = new();
+        builder(routingResponseBuilder);
+        return routingResponseBuilder.RoutingResponse;
+    }
     /// <summary>
     /// Entrypoint for extensions.
     /// </summary>