+++ src/TestableHttpClient/TestableHttpMessageHandlerOptions.cs
@@ -7,4 +7,6 @@ public sealed class TestableHttpMessageHandlerOptions
         DictionaryKeyPolicy = JsonNamingPolicy.CamelCase,
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase
     };
+
+    public RoutingOptions RoutingOptions { get; } = new RoutingOptions();
 }