+++ src/TestableHttpClient/Utils/RoutingResponseBuilder.cs
@@ -1,0 +1,11 @@
+﻿using TestableHttpClient.Response;
+
+namespace TestableHttpClient.Utils;
+
+internal class RoutingResponseBuilder : IRoutingResponseBuilder
+{
+    public RoutingResponse RoutingResponse { get; } = new();
+
+    public void MapFallBackResponse(IResponse fallBackResponse) => RoutingResponse.FallBackResponse = fallBackResponse;
+    public void Map(string route, IResponse response) => RoutingResponse.ResponseMap.Add(RouteParser.Parse(route), response);
+}